src/requirements.txt
asgiref==3.5.0
Django==4.0.2
django-audit-log==0.7.0
django-auditlog==1.0.0
django-bootstrap3==21.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-jsonfield-backport==1.0.5
django-tables2==2.4.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
mysqlclient==2.1.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2